# Conflicts:
#	frontend/src/app/app.module.ts

+++ frontend/src/app/components/workspace/key/key.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { ComplexStringService } from '../../../services/complex-string.service';
 import { MatDialog } from '@angular/material';
 import { ImgDialogComponent } from '../../../dialogs/img-dialog/img-dialog.component';
+import { ConfirmDialogComponent } from '../../../dialogs/confirm-dialog/confirm-dialog.component';
+import {SnotifyService, SnotifyPosition, SnotifyToastConfig} from 'ng-snotify';
 
 @Component({
   selector: 'app-workspace-key',
@@ -12,18 +14,37 @@ export class KeyComponent implements OnInit {
 
   @Input() public key: any;
   @Output() idEvent = new EventEmitter<number>();
-
+  description: string = "Are you sure you want to remove the string?";
+  btnOkText: string = "Delete";
+  btnCancelText: string = "Cancel";
+  answer: boolean;
 
   constructor(private dataProvider: ComplexStringService,
-              public dialog: MatDialog) { }
+              public dialog: MatDialog,
+              private snotifyService: SnotifyService) { }
 
   ngOnInit() {
   }
 
   onDeleteString() {
-    this.dataProvider.delete(this.key.id)
-    .subscribe(() => {});
-    this.idEvent.emit(this.key.id);
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      width: '500px',
+      data: {description: this.description, btnOkText: this.btnOkText, btnCancelText: this.btnCancelText, answer: this.answer}
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      if (dialogRef.componentInstance.data.answer){
+        this.dataProvider.delete(this.key.id)
+        .subscribe(
+          (response => {
+            this.snotifyService.success("String deleted", "Success!");
+          }),
+          err => {
+            this.snotifyService.error("String wasn`t deleted", "Error!");
+          });
+          this.idEvent.emit(this.key.id);
+        }
+      }
+    );
   }
 
   onPictureIconClick(key: any){
@@ -31,7 +52,6 @@ export class KeyComponent implements OnInit {
       data: {
         imageUri: key.pictureLink
       }
-      });
+    });
   }
-
 }
